Web_Programs/calculatorWithBackend2Deploy/calculator2Backend/calcbackend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.5
django-cors-headers==4.9.0
djangorestframework==3.17.1
gunicorn==26.0.0
packaging==26.2
python-decouple==3.8
sqlparse==0.5.5
whitenoise==6.12.0
